api-notes/requirements.txt
bottle==0.12.13
marshmallow==3.6.1
peewee==3.13.3
PyJWT==1.7.1
Marshmallow-Peewee==3.0.0 
bcrypt==3.1.7
bottle-cors==0.1.5